Refactor errors are caught, to support v12
Between v8 and v10, the 'PROMISE' type stopped exposing the actual
promise meaning that we couldn't do the hacky workaround for how it
drops the EID in those cases.

As a result, any unhandled promise rejection that wasn't handled by this
module turned into a fatal exception!  I figured this was just a new
node feature at first, and kind of liked it, but it _is_ a bug and an
overreach for a module that is intended to be less opinionated than
that.

This refactor handles all of the --unhandled-rejection arguments node
supports, while only calling the domain.onerror function once for a
given promise failure.  Works as expected in v8 and v10, but as they
lack support for these arguments, full coverage is impossible there.