Remove promiseExecutionId reset in after hook
The promiseExecutionId variable is a workaround for a weirdness in the
async_hooks API where promises didn't get cleaned up properly.

I was resetting it back to null in the 'after' hook for a Promise, but
node v13 appears to have changed the ordering of when 'after' events
happen relative to promise rejections, resulting in that value being
reset to null before the promise rejection handler fired.

There's no real reason to set it back to null, so I removed that code,
and now it works the same on node versions 8 - 13.